Update host_disks columns to use DECIMAL instead of FLOAT

+++ server/datastore/mysql/migrations/tables/20220914154915_CreateHostDisksTable.go
@@ -12,8 +12,8 @@ func Up_20220914154915(tx *sql.Tx) error {
 	_, err := tx.Exec(`
     CREATE TABLE host_disks (
         host_id                      INT(10) UNSIGNED NOT NULL,
-        gigs_disk_space_available    FLOAT NOT NULL DEFAULT 0,
-        percent_disk_space_available FLOAT NOT NULL DEFAULT 0,
+        gigs_disk_space_available    DECIMAL(10,2) NOT NULL DEFAULT 0,
+        percent_disk_space_available DECIMAL(10,2) NOT NULL DEFAULT 0,
         created_at                   TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
         updated_at                   TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
